requirements.txt
aiohttp==3.6.2
async-timeout==3.0.1
asyncio==3.4.3
attrs==19.3.0
cffi==1.13.2
chardet==3.0.4
discord.py==1.3.0
ffmpeg==1.4
idna==2.8
multidict==4.7.4
pycparser==2.19
PyNaCl==1.3.0
six==1.14.0
websockets==8.1
yarl==1.4.2
